requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.4
dataclasses-json==0.5.14
Django==4.2.5
django-colorfield==0.10.1
django-easy-select2==1.5.8
django-grappelli==3.0.8
django-maintenance-mode==0.18.0
django-multiupload==0.6.1
django-nested-admin==4.0.2
facepy==1.0.12
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
install==1.3.5
jwt==1.3.1
marshmallow==3.20.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycurl @ file:///C:/Users/Aigap/Downloads/pycurl-7.43.0.5-cp38-cp38-win_amd64.whl#sha256=1957c867e2a341f5526c107c7bbc5014d6e75fdc2a14294fcb8a47663fbd2e15
pyfacebook==0.3
pyparsing==3.1.1
python-facebook-api==0.17.1
python-fsutil==0.10.0
python-monkey-business==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
